+++ pybamm/models/submodels/convection/transverse/full_convection.py
@@ -39,8 +39,18 @@ class Full(BaseTransverseModel):
         return variables
 
     def set_algebraic(self, variables):
+        param = self.param
+
         p_s = variables["X-averaged separator pressure"]
-        div_Vbox_s = self._get_separator_velocity(variables)
+        # Difference in negative and positive electrode velocities determines the
+        # velocity in the separator
+        i_boundary_cc = variables["Current collector current density"]
+        v_box_n_right = param.beta_n * i_boundary_cc
+        v_box_p_left = param.beta_p * i_boundary_cc
+        d_vbox_s_dx = (v_box_p_left - v_box_n_right) / param.l_s
+
+        # Simple formula for velocity in the separator
+        div_Vbox_s = -d_vbox_s_dx
         Vbox_s = variables["X-averaged separator transverse volume-averaged velocity"]
 
         # Problem in the z-direction for p_s