+++ tests/unit/test_models/test_full_battery_models/test_lead_acid/test_basic_models.py
@@ -6,7 +6,7 @@ import unittest
 
 
 class TestBasicModels(unittest.TestCase):
-    def test_dfn_well_posed(self):
+    def test_basic_full_lead_acid_well_posed(self):
         model = pybamm.lead_acid.BasicFull()
         model.check_well_posedness()