+++ tests/unit/test_models/test_full_battery_models/test_lead_acid/test_loqs.py
@@ -26,7 +26,7 @@ class TestLeadAcidLOQS(unittest.TestCase):
         self.assertTrue(
             isinstance(
                 model.default_spatial_methods["current collector"],
-                pybamm.ZeroDimensionalMethod,
+                pybamm.ZeroDimensionalSpatialMethod,
             )
         )
         self.assertTrue(